@@ -2,14 +2,14 @@ import java.util.regex.Matcher
 
 plugins {
     id 'application'
-    id 'org.openjfx.javafxplugin' version '0.0.13'
-    id 'com.diffplug.spotless' version '6.19.0'
+    id 'org.openjfx.javafxplugin' version '0.0.14'
+    id 'com.diffplug.spotless' version '6.25.0'
 }
 
 group 'de.unijena.cheminf'
-version '1.1.1.0'
+version '1.2.0.0-beta'
 
-sourceCompatibility = 1.17
+sourceCompatibility = 1.21
 
 //Creates javadoc and sources jars
 java {
@@ -21,14 +21,17 @@ repositories {
     mavenCentral()
 }
 
-def javaFxVersion = '17.0.2'
-def cdkVersion = '2.8'
+def javaFxVersion = '21.0.1'
+def cdkVersion = '2.9'
 
 dependencies {
     implementation 'org.jetbrains:annotations:19.0.0'
     testImplementation(platform('org.junit:junit-bom:5.9.3'))
     testImplementation ('org.junit.jupiter:junit-jupiter')
     implementation group: 'org.openscience.cdk', name: 'cdk-bundle', version: cdkVersion
+    implementation group: 'org.openscience.cdk', name: 'cdk-scaffold', version: '2.8'
+    implementation group: 'io.github.jonasschaub', name: 'sru', version: '1.4.0.0'
+    implementation group: 'io.github.jonasschaub', name: 'ErtlFunctionalGroupsFinder', version: '1.3.0.0'
     implementation group: 'com.github.librepdf', name: 'openpdf', version: '1.3.26'
     implementation group: 'org.openjfx', name: 'javafx-controls', version: javaFxVersion, classifier: 'win'
     implementation group: 'org.openjfx', name: 'javafx-controls', version: javaFxVersion, classifier: 'linux'
@@ -73,7 +76,7 @@ startScripts {
     doLast {
         windowsScript.text = windowsScript.text.replaceAll('set CLASSPATH=.*', 'set CLASSPATH=.;%APP_HOME%/lib/*')
         windowsScript.text = windowsScript.text.replace("set JAVA_HOME=%JAVA_HOME:\"=%",
-                "set JAVA_HOME=%APP_HOME%\\jdk-17.0.4_8_jre\\")
+                "set JAVA_HOME=%APP_HOME%\\jdk-21.0.1_12_jre\\")
         windowsScript.text = windowsScript.text.replaceAll("java.exe", "javaw.exe")
         //windowsScript.text = windowsScript.text.replace("\"%JAVA_EXE%\" %DEFAULT_JVM_OPTS%", "start \"MORTAR\" \"%JAVA_EXE%\" %DEFAULT_JVM_OPTS%") //dirty workaround to suppress command window
         //dirty workaround to suppress command window and pass CMD options to MORTAR to suppress the Java version check
@@ -94,7 +97,7 @@ task updateHeapSpace20GbExtraRunApp(type: CreateStartScripts) {
         doLast {
             windowsScript.text = windowsScript.text.replaceAll('set CLASSPATH=.*', 'set CLASSPATH=.;%APP_HOME%/lib/*')
             windowsScript.text = windowsScript.text.replace("set JAVA_HOME=%JAVA_HOME:\"=%",
-                    "set JAVA_HOME=%APP_HOME%\\jdk-17.0.4_8_jre\\")
+                    "set JAVA_HOME=%APP_HOME%\\jdk-21.0.1_12_jre\\")
             windowsScript.text = windowsScript.text.replaceAll("java.exe", "javaw.exe")
             //windowsScript.text = windowsScript.text.replace("\"%JAVA_EXE%\" %DEFAULT_JVM_OPTS%", "start \"MORTAR\" \"%JAVA_EXE%\" %DEFAULT_JVM_OPTS%") //dirty workaround to suppress command window
             //dirty workaround to suppress command window and pass CMD options to MORTAR to suppress the Java version check
@@ -169,11 +172,6 @@ distributions {
 
 spotless {
     java {
-        target fileTree('.') {
-            include '**/*.java'
-            //exclude all classes that were just copied into this project
-            exclude '**/deglycosylation/**', '**/scaffoldGenerator/**', '**/cdk/**'
-        }
         toggleOffOn() //all code in between spotless:off / spotless:on will be ignored
         licenseHeaderFile('License-header/License-header.txt')
         encoding 'UTF-8'
