@@ -1,21 +1,26 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.model.fragmentation.algorithm;
@@ -47,7 +52,7 @@ public class ErtlFunctionalGroupsFinderFragmenterTest {
      * fragmenter because the settings tooltips are imported from the message.properties file.
      */
     public ErtlFunctionalGroupsFinderFragmenterTest() {
-        Locale.setDefault(new Locale("en", "GB"));
+        Locale.setDefault(Locale.of("en", "GB"));
     }
     //
     /**
@@ -58,16 +63,17 @@ public class ErtlFunctionalGroupsFinderFragmenterTest {
     @Test
     public void basicTest() throws Exception {
         ErtlFunctionalGroupsFinderFragmenter tmpFragmenter = new ErtlFunctionalGroupsFinderFragmenter();
-        System.out.println(tmpFragmenter.getFragmentationAlgorithmName());
-        System.out.println(tmpFragmenter.getElectronDonationModelSetting());
-        System.out.println(tmpFragmenter.getEnvironmentModeSetting());
-        for (Property tmpSetting : tmpFragmenter.settingsProperties()) {
-            System.out.println(tmpSetting.getName());
+        Assertions.assertDoesNotThrow(tmpFragmenter::getFragmentationAlgorithmName);
+        Assertions.assertDoesNotThrow(tmpFragmenter::getFragmentationAlgorithmDisplayName);
+        Assertions.assertDoesNotThrow(tmpFragmenter::getElectronDonationModelSetting);
+        Assertions.assertDoesNotThrow(tmpFragmenter::getEnvironmentModeSetting);
+        for (Property<?> tmpSetting : tmpFragmenter.settingsProperties()) {
+            Assertions.assertDoesNotThrow(tmpSetting::getName);
         }
     }
     //
     /**
-     * Does a test fragmentation on the COCONUT natural product CNP0151033 and prints the results.
+     * Does a test fragmentation on the COCONUT natural product CNP0151033.
      *
      * @throws Exception if anything goes wrong
      */
@@ -80,7 +86,7 @@ public class ErtlFunctionalGroupsFinderFragmenterTest {
         ErtlFunctionalGroupsFinderFragmenter tmpFragmenter = new ErtlFunctionalGroupsFinderFragmenter();
         tmpFragmenter.setEnvironmentModeSetting(ErtlFunctionalGroupsFinderFragmenter.FGEnvOption.NO_ENVIRONMENT);
         tmpFragmenter.setEnvironmentModeSetting(ErtlFunctionalGroupsFinderFragmenter.FGEnvOption.GENERALIZATION);
-        tmpFragmenter.environmentModeSettingProperty().set(ErtlFunctionalGroupsFinderFragmenter.FGEnvOption.FULL_ENVIRONMENT.name());
+        tmpFragmenter.environmentModeSettingProperty().set(ErtlFunctionalGroupsFinderFragmenter.FGEnvOption.FULL_ENVIRONMENT);
         tmpFragmenter.setFragmentSaturationSetting(IMoleculeFragmenter.FragmentSaturationOption.HYDROGEN_SATURATION);
         tmpFragmenter.setElectronDonationModelSetting(ErtlFunctionalGroupsFinderFragmenter.ElectronDonationModelOption.CDK);
         tmpFragmenter.setReturnedFragmentsSetting(ErtlFunctionalGroupsFinderFragmenter.EFGFFragmenterReturnedFragmentsOption.ALL_FRAGMENTS);
@@ -92,8 +98,8 @@ public class ErtlFunctionalGroupsFinderFragmenterTest {
         Assertions.assertTrue(tmpFragmenter.canBeFragmented(tmpOriginalMolecule));
         tmpFragmentList = tmpFragmenter.fragmentMolecule(tmpOriginalMolecule);
         for (IAtomContainer tmpFragment : tmpFragmentList) {
-            System.out.println(tmpSmiGen.create(tmpFragment) + " " + tmpFragment.getProperty(
-                    IMoleculeFragmenter.FRAGMENT_CATEGORY_PROPERTY_KEY));
+            Assertions.assertDoesNotThrow(() -> tmpSmiGen.create(tmpFragment));
+            Assertions.assertNotNull(tmpFragment.getProperty(IMoleculeFragmenter.FRAGMENT_CATEGORY_PROPERTY_KEY));
         }
     }
 }
