@@ -1,21 +1,26 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.gui.controls;
@@ -30,131 +35,129 @@ import javafx.scene.control.SeparatorMenuItem;
 
 /**
  * A MenuBar for the application's {@link MainView}.
- * It contains menus for file handling (I/O), shutting down the application, settings and help menu entries
+ * It contains menus for file handling (I/O), shutting down the application, settings and help menu entries.
  *
  * @author Felix Baensch, Jonas Schaub, Samuel Behr
  * @version 1.0.0.0
  */
 public class MainMenuBar extends MenuBar {
-
-    //<editor-fold desc="private class variables" defaultstate="collapsed">
+    //<editor-fold desc="private final class constants" defaultstate="collapsed">
     /**
-     * FileMenu
+     * FileMenu.
      */
-    private Menu fileMenu;
+    private final Menu fileMenu;
     /**
-     * MenuItem to import molecules
+     * MenuItem to import molecules.
      */
-    private MenuItem openMenuItem;
+    private final MenuItem openMenuItem;
     /**
-     * MenuItem for export
+     * MenuItem for export.
      */
-    private Menu exportMenu;
+    private final Menu exportMenu;
     /**
-     * Menu for fragments export
+     * Menu for fragments export.
      */
-    private Menu fragmentsExportMenu;
+    private final Menu fragmentsExportMenu;
     /**
-     * MenuItem to export fragments as csv file
+     * MenuItem to export fragments as csv file.
      */
-    private MenuItem fragmentsExportToCSVMenuItem;
+    private final MenuItem fragmentsExportToCSVMenuItem;
     /**
-     * MenuItem to export fragments as pdb file
+     * MenuItem to export fragments as pdb file.
      */
-    private MenuItem fragmentsExportToPDBMenuItem;
+    private final MenuItem fragmentsExportToPDBMenuItem;
     /**
-     * MenuItem to export fragments as pdf file
+     * MenuItem to export fragments as pdf file.
      */
-    private MenuItem fragmentsExportToPDFMenuItem;
+    private final MenuItem fragmentsExportToPDFMenuItem;
     /**
-     * Menu for fragments export as sdf
+     * Menu for fragments export as sdf.
      */
-    private Menu fragmentsExportToSDFMenu;
+    private final Menu fragmentsExportToSDFMenu;
     /**
-     * MenuItem to export fragments as sd file
+     * MenuItem to export fragments as sd file.
      */
-    private MenuItem fragmentsExportToSingleSDFMenuItem;
+    private final MenuItem fragmentsExportToSingleSDFMenuItem;
     /**
-     * MenuItem to export fragments as sd file separately
+     * MenuItem to export fragments as sd file separately.
      */
-    private MenuItem fragmentsExportToSeparateSDFsMenuItem;
+    private final MenuItem fragmentsExportToSeparateSDFsMenuItem;
     /**
-     * Menu for items export
+     * Menu for items export.
      */
-    private Menu itemsExportMenu;
+    private final Menu itemsExportMenu;
     /**
-     * MenuItem to export items as csv file
+     * MenuItem to export items as csv file.
      */
-    private MenuItem itemsExportToCSVMenuItem;
+    private final MenuItem itemsExportToCSVMenuItem;
     /**
-     * MenuItem to export fragments as pdf file
+     * MenuItem to export fragments as pdf file.
      */
-    private MenuItem itemsExportToPDFMenuItem;
+    private final MenuItem itemsExportToPDFMenuItem;
     /**
-     * MenuItem to exit app
+     * MenuItem to exit app.
      */
-    private MenuItem exitMenuItem;
+    private final MenuItem exitMenuItem;
     /**
-     * Menu for settings
+     * Menu for settings.
      */
-    private Menu settingsMenu;
+    private final Menu settingsMenu;
     /**
-     * Menu for pipeline
+     * Menu for pipeline.
      */
-    private Menu pipelineMenu;
+    private final Menu pipelineMenu;
     /**
-     * Menu for help
+     * Menu for help.
      */
-    private Menu helpMenu;
+    private final Menu helpMenu;
     /**
-     * MenuItem to open fragmentation settings
+     * MenuItem to open fragmentation settings.
      */
-    private MenuItem fragmentationSettingsMenuItem;
+    private final MenuItem fragmentationSettingsMenuItem;
     /**
-     * MenuItem to open global settings
+     * MenuItem to open global settings.
      */
-    private MenuItem globalSettingsMenuItem;
+    private final MenuItem globalSettingsMenuItem;
     /**
-     * MenuItem to open pipeline settings
+     * MenuItem to open pipeline settings.
      */
-    private MenuItem pipelineSettingsMenuItem;
+    private final MenuItem pipelineSettingsMenuItem;
     /**
-     * MenuItem to open AboutView
+     * MenuItem to open AboutView.
      */
-    private MenuItem aboutViewMenuItem;
+    private final MenuItem aboutViewMenuItem;
     /**
-     * Menu to choose fragmentation algorithm
+     * Menu to choose fragmentation algorithm.
      */
-    private Menu fragmentationAlgorithmMenu;
+    private final Menu fragmentationAlgorithmMenu;
     /**
-     * MenuItem to cancel molecule import, only visible if import is running
+     * MenuItem to cancel molecule import, only visible if import is running.
      */
-    private MenuItem cancelImportMenuItem;
+    private final MenuItem cancelImportMenuItem;
     /**
-     * MenuItem to cancel export, only visible if import is running
+     * MenuItem to cancel export, only visible if import is running.
      */
-    private MenuItem cancelExportMenuItem;
+    private final MenuItem cancelExportMenuItem;
     /**
-     * Menu to open views
+     * Menu to open views.
      */
-    private Menu viewsMenu;
+    private final Menu viewsMenu;
     /**
-     * MenuItem to open the histogram
+     * MenuItem to open the histogram.
      */
-    private MenuItem histogramViewerMenuItem;
+    private final MenuItem histogramViewerMenuItem;
     /**
-     * MenuItem to open the OverviewView
+     * MenuItem to open the OverviewView.
      */
-    private MenuItem overviewViewMenuItem;
+    private final MenuItem overviewViewMenuItem;
     //</editor-fold>
     //
     //<editor-fold desc="constructor" defaultstate="collapsed">
     /**
-     * Constructor
-     * It initialises the MainMenuBar's components, adds them to the MainMenuBar.
-     * No listeners were added here.
+     * Initialises the MainMenuBar's components and adds them to the MainMenuBar.
+     * No listeners are added here.
      */
-    public MainMenuBar(){
+    public MainMenuBar() {
         super();
         //<editor-fold desc="initialisation" defaultstate="collapsed">
         //fileMenu
@@ -205,9 +208,9 @@ public class MainMenuBar extends MenuBar {
     //<editor-fold desc="private methods" defaultstate="collapsed">
     //<editor-fold desc="addComponentsToMenuBar" defaultstate="collapsed">
     /**
-     * Adds the menu items to the menus and the menus to the menu bar
+     * Adds the menu items to the menus and the menus to the menu bar.
      */
-    private void addComponentsToMenuBar(){
+    private void addComponentsToMenuBar() {
         //<editor-fold desc="fileMenu" defaultstate="collapsed">
         this.getMenus().add(this.fileMenu);
         //openMenuItem
@@ -259,9 +262,9 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="addComponentsToExportMenu" defaultstate="collapsed">
     /**
-     * Adds the menu items to the export menu
+     * Adds the menu items to the export menu.
      */
-    private void addComponentsToExportMenu(){
+    private void addComponentsToExportMenu() {
         //<editor-fold desc="fragmentsExportMenu" defaultstate="collapsed">
         this.exportMenu.getItems().add(this.fragmentsExportMenu);
         this.fragmentsExportMenu.getItems().add(this.fragmentsExportToCSVMenuItem);
@@ -270,7 +273,6 @@ public class MainMenuBar extends MenuBar {
         this.fragmentsExportMenu.getItems().add(this.fragmentsExportToSDFMenu);
         this.fragmentsExportToSDFMenu.getItems().add(this.fragmentsExportToSingleSDFMenuItem);
         this.fragmentsExportToSDFMenu.getItems().add(this.fragmentsExportToSeparateSDFsMenuItem);
-//        this.exportMenu.show();
         //</editor-fold>
         //<editor-fold desc="itemsExportMenu" defaultstate="collapsed">
         this.exportMenu.getItems().add(this.itemsExportMenu);
@@ -284,7 +286,7 @@ public class MainMenuBar extends MenuBar {
     //<editor-fold desc="public properties" defaultstate="collapsed">
     //<editor-fold desc="getOpenMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open a molecule set
+     * Returns the menu item that is supposed to open a molecule set.
      *
      * @return the menu item that should open a molecule set
      */
@@ -294,7 +296,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getExportMenu" defaultstate="collapsed">
     /**
-     * Returns the menu that is supposed to open a list of export options
+     * Returns the menu that is supposed to open a list of export options.
      *
      * @return the menu that should open a list of export options
      */
@@ -304,7 +306,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportMenu" defaultstate="collapsed">
     /**
-     * Returns the menu that is supposed to open a list of fragments export options
+     * Returns the menu that is supposed to open a list of fragments export options.
      *
      * @return the menu that should open a list of fragments export options
      */
@@ -314,7 +316,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToCSVMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the fragments to a CSV file
+     * Returns the menu item that is supposed to export the fragments to a CSV file.
      *
      * @return the menu item that should export the fragments to a CSV file
      */
@@ -324,7 +326,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToPDBMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the fragments to a PDB file
+     * Returns the menu item that is supposed to export the fragments to a PDB file.
      *
      * @return the menu item that should export the fragments to a PDB file
      */
@@ -334,7 +336,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToPDFMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the fragments to a PDF
+     * Returns the menu item that is supposed to export the fragments to a PDF.
      *
      * @return the menu item that should export the fragments to a PDF
      */
@@ -344,7 +346,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToSDFMenu" defaultstate="collapsed">
     /**
-     * Returns the menu that is supposed to open a list of SDF export options
+     * Returns the menu that is supposed to open a list of SDF export options.
      *
      * @return the menu that should open a list of SDF export options
      */
@@ -354,7 +356,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToSingleSDFMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the fragments to a single SD file
+     * Returns the menu item that is supposed to export the fragments to a single SD file.
      *
      * @return the menu item that should export the fragments to a single SD file
      */
@@ -364,7 +366,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentsExportToSeparateSDFsMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the fragments to separate SD files
+     * Returns the menu item that is supposed to export the fragments to separate SD files.
      *
      * @return the menu item that should export the fragments to separate SD files
      */
@@ -374,7 +376,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getItemsExportMenu" defaultstate="collapsed">
     /**
-     * Returns the menu that is supposed to open a list of items export options
+     * Returns the menu that is supposed to open a list of items export options.
      *
      * @return the menu that should open a list of items export options
      */
@@ -384,7 +386,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getItemsExportToCSVMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the items to a CSV file
+     * Returns the menu item that is supposed to export the items to a CSV file.
      *
      * @return the menu item that should export the items to a CSV file
      */
@@ -394,7 +396,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getItemsExportToPDFMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to export the items to a PDF
+     * Returns the menu item that is supposed to export the items to a PDF.
      *
      * @return the menu item that should export the items to a PDF
      */
@@ -404,7 +406,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getExitMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to shut down the application
+     * Returns the menu item that is supposed to shut down the application.
      *
      * @return the menu item that should shut down the application
      */
@@ -414,7 +416,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getGlobalSettingsMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that opens the global settings view
+     * Returns the menu item that opens the global settings view.
      *
      * @return the menu item that opens the global settings view
      */
@@ -424,7 +426,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentationAlgorithmMenu" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to choose the fragmentation algorithm
+     * Returns the menu item that is supposed to choose the fragmentation algorithm.
      *
      * @return the menu item that should choose the fragmentation algorithm
      */
@@ -434,7 +436,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getFragmentationSettingsMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open the fragmentation settings window
+     * Returns the menu item that is supposed to open the fragmentation settings window.
      *
      * @return the menu item that should open the fragmentation settings window
      */
@@ -444,7 +446,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getPipelineSettingsMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open the pipeline settings window
+     * Returns the menu item that is supposed to open the pipeline settings window.
      *
      * @return the menu item that should open the pipeline settings window
      */
@@ -454,7 +456,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getOverviewMenuItem defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open the OverviewView
+     * Returns the menu item that is supposed to open the OverviewView.
      *
      * @return the menu item that should open the OverviewView
      */
@@ -464,7 +466,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getAboutViewMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open the AboutView
+     * Returns the menu item that is supposed to open the AboutView.
      *
      * @return the menu item that should open the AboutView
      */
@@ -474,7 +476,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //<editor-fold desc="getHistogramViewerMenuItem" defaultstate="collapsed">
     /**
-     * Returns the menu item that is supposed to open histogram window
+     * Returns the menu item that is supposed to open histogram window.
      *
      * @return MenuItem to open histogram view
      */
@@ -484,7 +486,7 @@ public class MainMenuBar extends MenuBar {
     //</editor-fold>
     //
     /**
-     * Returns MenuItem to cancel running import, only visible if import is running
+     * Returns MenuItem to cancel running import, only visible if import is running.
      *
      * @return MenuItem to cancel import
      */
@@ -493,7 +495,7 @@ public class MainMenuBar extends MenuBar {
     }
     //
     /**
-     * Returns MenuItem to cancel running export, only visible if import is running
+     * Returns MenuItem to cancel running export, only visible if import is running.
      *
      * @return MenuItem to cancel export
      */
