@@ -1,21 +1,26 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.gui.controls;
@@ -30,26 +35,25 @@ import javafx.scene.control.ProgressBar;
 import javafx.scene.layout.FlowPane;
 
 /**
- * StatusBar to show status of the application and progress of a running task
+ * StatusBar to show status of the application and progress of a running task.
  *
  * @author Felix Baensch, Jonas Schaub
  * @version 1.0.0.0
  */
 public class StatusBar extends FlowPane {
-
-    //<editor-fold desc="private class variables" defaultstate="collapsed">
+    //<editor-fold desc="private final class constants" defaultstate="collapsed">
     /**
-     * Label to show status message
+     * Label to show status message.
      */
-    private Label statusLabel;
+    private final Label statusLabel;
     /**
-     * ProgressBar
+     * ProgressBar.
      */
-    private ProgressBar progressBar;
+    private final ProgressBar progressBar;
     //</editor-fold>
     //
     /**
-     * Constructor
+     * Constructor.
      */
     public StatusBar(){
         super();
@@ -68,7 +72,8 @@ public class StatusBar extends FlowPane {
     //
     //<editor-fold desc="properties" defaultstate="collapsed">
     /**
-     * Returns statusLabel
+     * Returns statusLabel.
+     *
      * @return Label
      */
     public Label getStatusLabel() {
@@ -76,7 +81,8 @@ public class StatusBar extends FlowPane {
     }
     //
     /**
-     * Returns the progressBar
+     * Returns the progressBar.
+     *
      * @return ProgressBar
      */
     public ProgressBar getProgressBar() {
