@@ -1,61 +1,65 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.gui.util;
 
 /**
- * Wrapper class to visualize information about external tools used in MORTAR in the AboutView
+ * Wrapper class to visualize information about external tools used in MORTAR in the AboutView.
  *
  * @author Felix Baensch
  * @version 1.0.0.0
  */
 public class ExternalTool {
-
     //<editor-fold desc="private class variables" defaultstate="collapsed">
     /**
-     * Name of the tool
+     * Name of the tool.
      */
     private String name;
     /**
-     * Number string of the used version of the tool
+     * Number string of the used version of the tool.
      */
     private String version;
     /**
-     * Author(s) of the tool
+     * Author(s) of the tool.
      */
     private String author;
     /**
-     * License of the tool
+     * License of the tool.
      */
     private String license;
     //</editor-fold>
     //
     /**
-     * Constructor
+     * Constructor.
      *
      * @param aName Name of tool
      * @param aVersion Version used
      * @param anAuthor Author
      * @param aLicense License used
      */
-    public ExternalTool(String aName, String aVersion, String anAuthor, String aLicense){
+    public ExternalTool(String aName, String aVersion, String anAuthor, String aLicense) {
         this.name = aName;
         this.version = aVersion;
         this.author = anAuthor;
@@ -64,7 +68,7 @@ public class ExternalTool {
     //
     //<editor-fold desc="public properties" defaultstate="collapsed">
     /**
-     * Returns name as String
+     * Returns name as String.
      *
      * @return {@link String}
      */
@@ -73,7 +77,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Returns version number as String
+     * Returns version number as String.
      *
      * @return {@link String}
      */
@@ -82,7 +86,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Returns author(s) name
+     * Returns author(s) name.
      *
      * @return {@link String}
      */
@@ -91,7 +95,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Returns license
+     * Returns license.
      *
      * @return {@link String}
      */
@@ -100,7 +104,8 @@ public class ExternalTool {
     }
     //
     /**
-     * Sets given String as name
+     * Sets given String as name.
+     *
      * @param name {@link String}
      */
     public void setName(String name) {
@@ -108,7 +113,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Sets given String as version
+     * Sets given String as version.
      *
      * @param version {@link String}
      */
@@ -117,7 +122,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Sets given String as author
+     * Sets given String as author.
      *
      * @param author {@link String}
      */
@@ -126,7 +131,7 @@ public class ExternalTool {
     }
     //
     /**
-     * Sets given String as license
+     * Sets given String as license.
      *
      * @param license {@link String}
      */
