@@ -1,113 +1,128 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.model.util;
 
+import de.unijena.cheminf.mortar.model.data.DataModelPropertiesForTableView;
 import de.unijena.cheminf.mortar.model.data.FragmentDataModel;
 import de.unijena.cheminf.mortar.model.data.MoleculeDataModel;
 
-import java.util.Collections;
 import java.util.List;
 
 /**
  * Util class for collections.
  *
  * @author Felix Baensch, Jonas Schaub
- * @version 1.0.1.0
+ * @version 1.0.2.0
  */
 public final class CollectionUtil {
-    //<editor-fold desc="public static methods" defaultstate="collapsed">
-    //TODO: check parameters, change params to enum values (and bools?), throw exceptions!
+    //<editor-fold desc="Private constructor" defaultstate="collapsed">
     /**
-     * Sorts given list by property and sort type.
-     *
-     * @param aList List
-     * @param aProperty String
-     * @param aSortType String
+     * Private parameter-less constructor.
+     * Introduced because javadoc build complained about classes without declared default constructor.
      */
-    public static void sortGivenFragmentListByPropertyAndSortType(List<? extends MoleculeDataModel> aList, String aProperty, String aSortType) {
-        Collections.sort(aList, (m1, m2) -> {
+    private CollectionUtil() {
+    }
+    //</editor-fold>
+    //
+    //<editor-fold desc="public static methods" defaultstate="collapsed">
+    /**
+     * Sorts a given list of Molecule-/FragmentDataModel objects according to the specified property and
+     * the sort type specified by a Boolean. If true, the list is sorted in ascending order,
+     * otherwise in descending order.
+     *
+     * @param aList List of Molecule-/FragmentDataModel objects to be sorted.
+     * @param aProperty String property of Molecule-/FragmentDataModel by which to sort the list. Property must be part
+     *                  of {@link DataModelPropertiesForTableView}. If the property is part of
+     *                  {@link DataModelPropertiesForTableView} but not handled in this method the list will not be sorted.
+     * @param ascending {@code true} to sort the list in ascending order, {@code false} for descending order.
+     * @throws IllegalArgumentException If the specified property is not part of the Molecule-/FragmentDataModel
+     * properties displayed in the TableViews. Or if the specified property does not correspond to the specified
+     * Molecule-/FragmentDataModel object.
+     */
+    public static void sortGivenFragmentListByPropertyAndSortType(List<? extends MoleculeDataModel> aList, String aProperty, boolean ascending) {
+        aList.sort((m1, m2) -> {
             FragmentDataModel f1;
             FragmentDataModel f2;
-            switch (aProperty) {
-                case "absoluteFrequency":
-                    f1 = (FragmentDataModel) m1;
-                    f2 = (FragmentDataModel) m2;
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return (Integer.compare(f1.getAbsoluteFrequency(), f2.getAbsoluteFrequency()));
-                        case "DESCENDING":
-                            return (f1.getAbsoluteFrequency() > f2.getAbsoluteFrequency() ? -1 : (f1.getAbsoluteFrequency() == f2.getAbsoluteFrequency() ? 0 : 1));
-                    }
-                case "absolutePercentage":
-                    f1 = (FragmentDataModel) m1;
-                    f2 = (FragmentDataModel) m2;
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return (Double.compare(f1.getAbsolutePercentage(), f2.getAbsolutePercentage()));
-                        case "DESCENDING":
-                            return (f1.getAbsolutePercentage() > f2.getAbsolutePercentage() ? -1 : (f1.getAbsolutePercentage() == f2.getAbsolutePercentage() ? 0 : 1));
-                    }
-                case "moleculeFrequency":
-                    f1 = (FragmentDataModel) m1;
-                    f2 = (FragmentDataModel) m2;
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return (Double.compare(f1.getMoleculeFrequency(), f2.getMoleculeFrequency()));
-                        case "DESCENDING":
-                            return (f1.getMoleculeFrequency() > f2.getMoleculeFrequency() ? -1 : (f1.getMoleculeFrequency() == f2.getMoleculeFrequency() ? 0 : 1));
-                    }
-                case "moleculePercentage":
-                    f1 = (FragmentDataModel) m1;
-                    f2 = (FragmentDataModel) m2;
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return (Double.compare(f1.getMoleculePercentage(), f2.getMoleculePercentage()));
-                        case "DESCENDING":
-                            return (f1.getMoleculePercentage() > f2.getMoleculePercentage() ? -1 : (f1.getMoleculePercentage() == f2.getMoleculePercentage() ? 0 : 1));
-                    }
-                case "name":
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return m1.getName().compareTo(m2.getName());
-                        case "DESCENDING":
-                            return m2.getName().compareTo(m1.getName());
-                    }
-                case "uniqueSmiles":
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return m1.getUniqueSmiles().compareTo(m2.getUniqueSmiles());
-                        case "DESCENDING":
-                            return m2.getUniqueSmiles().compareTo(m1.getUniqueSmiles());
-                    }
-                case "parentMoleculeName":
-                    f1 = (FragmentDataModel) m1;
-                    f2 = (FragmentDataModel) m2;
-                    switch (aSortType) {
-                        case "ASCENDING":
-                            return f1.getParentMoleculeName().compareTo(f2.getParentMoleculeName());
-                        case "DESCENDING":
-                            return f2.getParentMoleculeName().compareTo(f1.getParentMoleculeName());
-                    }
+            DataModelPropertiesForTableView property = DataModelPropertiesForTableView.fromString(aProperty);
+            if (property == null) {
+                throw new IllegalArgumentException("Property is not part of the data model properties that are displayed in the TableViews: " + aProperty);
             }
-            return 0;
+            try {
+                switch (property) {
+                    case DataModelPropertiesForTableView.ABSOLUTE_FREQUENCY:
+                        f1 = (FragmentDataModel) m1;
+                        f2 = (FragmentDataModel) m2;
+                        if (ascending)
+                            return Integer.compare(f1.getAbsoluteFrequency(), f2.getAbsoluteFrequency());
+                        else
+                            return Integer.compare(f2.getAbsoluteFrequency(), f1.getAbsoluteFrequency());
+                    case DataModelPropertiesForTableView.ABSOLUTE_PERCENTAGE:
+                        f1 = (FragmentDataModel) m1;
+                        f2 = (FragmentDataModel) m2;
+                        if (ascending)
+                            return Double.compare(f1.getAbsolutePercentage(), f2.getAbsolutePercentage());
+                        else
+                            return Double.compare(f2.getAbsolutePercentage(), f1.getAbsolutePercentage());
+                    case DataModelPropertiesForTableView.MOLECULE_FREQUENCY:
+                        f1 = (FragmentDataModel) m1;
+                        f2 = (FragmentDataModel) m2;
+                        if (ascending)
+                            return Double.compare(f1.getMoleculeFrequency(), f2.getMoleculeFrequency());
+                        else
+                            return Integer.compare(f2.getMoleculeFrequency(), f1.getMoleculeFrequency());
+                    case DataModelPropertiesForTableView.MOLECULE_PERCENTAGE:
+                        f1 = (FragmentDataModel) m1;
+                        f2 = (FragmentDataModel) m2;
+                        if (ascending)
+                            return Double.compare(f1.getMoleculePercentage(), f2.getMoleculePercentage());
+                        else
+                            return Double.compare(f2.getMoleculePercentage(), f1.getMoleculePercentage());
+                    case DataModelPropertiesForTableView.NAME:
+                        if (ascending)
+                            return m1.getName().compareTo(m2.getName());
+                        else
+                            return m2.getName().compareTo(m1.getName());
+                    case DataModelPropertiesForTableView.UNIQUE_SMILES:
+                        if (ascending)
+                            return m1.getUniqueSmiles().compareTo(m2.getUniqueSmiles());
+                        else
+                            return m2.getUniqueSmiles().compareTo(m1.getUniqueSmiles());
+                    case DataModelPropertiesForTableView.PARENT_MOLECULE_NAME:
+                        f1 = (FragmentDataModel) m1;
+                        f2 = (FragmentDataModel) m2;
+                        if (ascending)
+                            return f1.getParentMoleculeName().compareTo(f2.getParentMoleculeName());
+                        else
+                            return f2.getParentMoleculeName().compareTo(f1.getParentMoleculeName());
+                    default:
+                        return 0;
+                }
+            } catch (ClassCastException anException) {
+                throw new IllegalArgumentException("The specified property does not match the specified data model object.");
+            }
         });
     }
     //
@@ -133,7 +148,7 @@ public final class CollectionUtil {
         if (aLoadFactor <= 0 || aLoadFactor > 1.0f) {
             throw new IllegalArgumentException("Load factor must be higher than 0 and not bigger than 1.0 but is " + aLoadFactor);
         }
-        float tmpInitialSize = (float) aNumberOfElements * (1.0f / aLoadFactor) + 2.0f;
+        float tmpInitialSize = aNumberOfElements * (1.0f / aLoadFactor) + 2.0f;
         return (int) tmpInitialSize;
     }
     //
