@@ -1,26 +1,36 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.model.util;
 
+import de.unijena.cheminf.mortar.configuration.Configuration;
+import de.unijena.cheminf.mortar.configuration.IConfiguration;
+
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -29,16 +39,15 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
- * File utility
+ * File utility.
  *
  * @author Achim Zielesny, Jonas Schaub, Felix Baensch
  * @version 1.0.0.0
  */
 public final class FileUtil {
-
     //<editor-fold desc="Private static class variables" defaultstate="collapsed">
     /**
-     * Cache String for app dir path
+     * Cache String for app dir path.
      */
     private static String appDirPath = null;
     //</editor-fold>
@@ -48,6 +57,29 @@ public final class FileUtil {
      * Logger of this class.
      */
     private static final Logger LOGGER = Logger.getLogger(FileUtil.class.getName());
+    //
+    /**
+     * Configuration class to read resource file paths from.
+     */
+    private static final IConfiguration CONFIGURATION;
+    static {
+        try {
+            CONFIGURATION = Configuration.getInstance();
+        } catch (IOException anIOException) {
+            //when MORTAR is run via MainApp.start(), the correct initialization of Configuration is checked there before
+            // FileUtil is accessed and this static initializer called
+            throw new NullPointerException("Configuration could not be initialized");
+        }
+    }
+    //</editor-fold>
+    //
+    //<editor-fold desc="Private constructor" defaultstate="collapsed">
+    /**
+     * Private parameter-less constructor.
+     * Introduced because javadoc build complained about classes without declared default constructor.
+     */
+    private FileUtil() {
+    }
     //</editor-fold>
     //
     //<editor-fold defaultstate="collapsed" desc="Public static methods">
@@ -80,7 +112,7 @@ public final class FileUtil {
     /**
      * Returns the name of the file without the file extension.
      *
-     * @param aFile whose name should ne return without extension
+     * @param aFile whose name should be returned without extension
      * @return file name without extension
      * @throws NullPointerException if given file is 'null'
      */
@@ -88,11 +120,11 @@ public final class FileUtil {
         //<editor-fold defaultstate="collapsed" desc="Checks">
         Objects.requireNonNull(aFile, "Given file is 'null'.");
         //</editor-fold>
-        return aFile.getName().replaceFirst("[.][^.]+$", ""); //cuts the
+        return aFile.getName().replaceFirst("[.][^.]+$", "");
     }
 
     /**
-     * Deletes single file
+     * Deletes single file.
      *
      * @param aFilePathname Full pathname of file to be deleted (may be null
      * then false is returned)
@@ -100,19 +132,18 @@ public final class FileUtil {
      */
     public static boolean deleteSingleFile(String aFilePathname) {
         // <editor-fold defaultstate="collapsed" desc="Checks">
-        if (aFilePathname == null ||
-                aFilePathname.isEmpty()
-        ) {
+        if (aFilePathname == null || aFilePathname.isEmpty()) {
             return false;
         }
         // </editor-fold>
         try {
             File tmpFile = new File(aFilePathname);
-            if (!tmpFile.isFile()) {
-                return true;
-            } else {
-                return tmpFile.delete();
+            if (tmpFile.isFile()) {
+                Files.delete(tmpFile.toPath());
             }
+            //if it is not a file, it does not exist and therefore must not be deleted
+            //if delete fails, it goes to catch, logs the detailed exception and returns false
+            return true;
         } catch (Exception anException) {
             FileUtil.LOGGER.log(Level.SEVERE, anException.toString(), anException);
             return false;
@@ -134,30 +165,27 @@ public final class FileUtil {
             return false;
         }
         // </editor-fold>
+        boolean tmpAllFilesDeletedSuccessfully = true;
         try {
             File tmpDirectory = new File(aDirectoryPath);
             if (!tmpDirectory.isDirectory()) {
                 return false;
             }
             File[] tmpFilesArray = tmpDirectory.listFiles();
-            boolean tmpAllFilesDeletedSuccessfully = true;
             for (File tmpFile : tmpFilesArray) {
                 if (tmpFile.isFile()) {
-                    boolean tmpFileDeleted = tmpFile.delete();
-                    if (!tmpFileDeleted) {
-                        tmpAllFilesDeletedSuccessfully = false;
-                    }
+                    Files.delete(tmpFile.toPath());
                 }
             }
-            return tmpAllFilesDeletedSuccessfully;
         } catch (Exception anException) {
             FileUtil.LOGGER.log(Level.SEVERE, anException.toString(), anException);
-            return false;
+            tmpAllFilesDeletedSuccessfully = false;
         }
+        return tmpAllFilesDeletedSuccessfully;
     }
 
     /**
-     * Creates directory and all non-existent ancestor directories if necessary
+     * Creates directory and all non-existent ancestor directories if necessary.
      *
      * @param aDirectoryPath Full directory path to be created
      * @return true: Directory already existed or was successfully created,
@@ -184,7 +212,7 @@ public final class FileUtil {
     }
 
     /**
-     * Creates empty file
+     * Creates empty file.
      *
      * @param aFilePathname Full file pathname
      * @return True: Empty file was created, false: Otherwise
@@ -215,7 +243,7 @@ public final class FileUtil {
      * path cannot be determined or data directory cannot be created
      */
     public static String getAppDirPath() throws SecurityException {
-        if(FileUtil.appDirPath != null){
+        if (FileUtil.appDirPath != null) {
             return FileUtil.appDirPath;
         }
         String tmpAppDir;
@@ -233,10 +261,10 @@ public final class FileUtil {
             throw new SecurityException("AppData (Windows) or user home directory path " + tmpAppDir + " is either no directory or does not exist.");
         if (tmpOS.contains("MAC"))
             tmpAppDir += File.separator + "Library" + File.separator + "Application Support";
-        tmpAppDir += File.separator + BasicDefinitions.MORTAR_VENDOR + File.separator + BasicDefinitions.MORTAR_DATA_DIRECTORY;
+        tmpAppDir += File.separator + FileUtil.CONFIGURATION.getProperty("mortar.vendor.name") + File.separator + FileUtil.CONFIGURATION.getProperty("mortar.dataDirectory.name");
         tmpAppDirFile = new File(tmpAppDir);
         boolean tmpSuccessful = true;
-        if(!tmpAppDirFile.exists())
+        if (!tmpAppDirFile.exists())
             tmpSuccessful = tmpAppDirFile.mkdirs();
         if (!tmpSuccessful)
             throw new SecurityException("Unable to create application data directory");
@@ -252,20 +280,23 @@ public final class FileUtil {
      * path cannot be determined or data directory cannot be created
      */
     public static String getSettingsDirPath() throws SecurityException {
-        return FileUtil.getAppDirPath() + File.separator + BasicDefinitions.SETTINGS_CONTAINER_FILE_DIRECTORY + File.separator;
+        return FileUtil.getAppDirPath() + File.separator + FileUtil.CONFIGURATION.getProperty("mortar.settingsDirectory.name") + File.separator;
     }
 
     /**
      * Returns a timestamp to add to a filename.
      *
-     * @throws DateTimeException if the time cannot be determined or formatted
-     * @return timestamp filename extension
+     * @return timestamp filename extension or a placeholder string if time stamp creation failed
      */
     public static String getTimeStampFileNameExtension() throws DateTimeException {
-        LocalDateTime tmpDateTime = LocalDateTime.now();
-        String tmpDateTimeAddition = tmpDateTime.format(DateTimeFormatter.ofPattern(
-                BasicDefinitions.FILENAME_TIMESTAMP_FORMAT));
-        return tmpDateTimeAddition;
+        try {
+            LocalDateTime tmpDateTime = LocalDateTime.now();
+            String tmpDateTimeAddition = tmpDateTime.format(DateTimeFormatter.ofPattern(BasicDefinitions.FILENAME_TIMESTAMP_FORMAT));
+            return tmpDateTimeAddition;
+        } catch (Exception e) {
+            FileUtil.LOGGER.log(Level.WARNING, e.toString(), e);
+            return BasicDefinitions.FILENAME_TIMESTAMP_FORMAT;
+        }
     }
 
     /**
@@ -286,16 +317,15 @@ public final class FileUtil {
         if (tmpLastChar == File.separatorChar)
             throw new IllegalArgumentException("Given file path is a directory.");
         //</editor-fold>
-        String tmpFilePath = aFilePath;
         String tmpFileExtension = aFileExtension;
         if (Objects.isNull(tmpFileExtension)) {
             tmpFileExtension = "";
         }
         int tmpFilesInThisMinuteCounter = 1;
-        File tmpFile = new File(tmpFilePath+ tmpFileExtension);
+        File tmpFile = new File(aFilePath + tmpFileExtension);
         if (tmpFile.exists()) {
-            while (tmpFilesInThisMinuteCounter <= Integer.MAX_VALUE) {
-                tmpFile = new File(tmpFilePath + "(" + tmpFilesInThisMinuteCounter + ")" + tmpFileExtension);
+            while (true) {
+                tmpFile = new File(aFilePath + "(" + tmpFilesInThisMinuteCounter + ")" + tmpFileExtension);
                 if (!tmpFile.exists()) {
                     break;
                 }
@@ -307,7 +337,35 @@ public final class FileUtil {
             String tmpNonExistingFilePath = tmpFile.getPath();
             return tmpNonExistingFilePath;
         } else {
-            return tmpFilePath + tmpFileExtension;
+            return aFilePath + tmpFileExtension;
+        }
+    }
+    //
+    /**
+     * Opens given path in OS depending explorer equivalent.
+     *
+     * @param aPath path to open
+     * @throws IllegalArgumentException if the given path is empty, blank, or null
+     * @throws SecurityException if the directory could not be opened
+     */
+    public static void openFilePathInExplorer(String aPath) throws SecurityException {
+        if (Objects.isNull(aPath) || aPath.isEmpty() || aPath.isBlank()) {
+            throw new IllegalArgumentException("Given file path is null or empty.");
+        }
+        String tmpOS = System.getProperty("os.name").toUpperCase();
+        try {
+            if (tmpOS.contains("WIN")) {
+                Runtime.getRuntime().exec(new String[]{"explorer", "/open,", aPath});
+            } else if (tmpOS.contains("MAC")) {
+                Runtime.getRuntime().exec(new String[]{"open", "-R", aPath});
+            } else if (tmpOS.contains("NUX") || tmpOS.contains("NIX") || tmpOS.contains("AIX")) {
+                Runtime.getRuntime().exec(new String[]{"gio", "open", aPath});
+            } else {
+                throw new SecurityException("OS name " + tmpOS + " unknown.");
+            }
+        } catch (IOException anException) {
+            FileUtil.LOGGER.log(Level.SEVERE, anException.toString(), anException);
+            throw new SecurityException("Could not open directory path");
         }
     }
     // </editor-fold>
