@@ -1,27 +1,31 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.model.util;
 
 import java.util.Objects;
-import java.util.logging.Logger;
 
 /**
  * A generic sort wrapper that can be compared to another object (e.g. for sorting) via a specified sort string.
@@ -30,13 +34,8 @@ import java.util.logging.Logger;
  * @author Jonas Schaub
  * @version 1.0.0.0
  */
-public class StringSortWrapper<T> implements Comparable<StringSortWrapper>{
+public class StringSortWrapper<T> implements Comparable<StringSortWrapper<T>>{
     //<editor-fold defaultstate="collapsed" desc="Private static final class constants">
-    /**
-     * Logger of this class.
-     */
-    private static final Logger LOGGER = Logger.getLogger(StringSortWrapper.class.getName());
-
     /**
      * Seed for hashCode() method.
      */
@@ -135,7 +134,7 @@ public class StringSortWrapper<T> implements Comparable<StringSortWrapper>{
         if (anObject == null || (anObject.getClass() != this.getClass())) {
             return false;
         }
-        StringSortWrapper tmpStringSortWrapper = (StringSortWrapper) anObject;
+        StringSortWrapper<T> tmpStringSortWrapper = (StringSortWrapper<T>) anObject;
         return this.hashCode() == tmpStringSortWrapper.hashCode();
     }
 
