@@ -1,30 +1,29 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.preference;
-
-/**
- * TODO:
- * - Implement preference container reference (GUID) to ensure that no preference is added to multiple preference
- *      containers?
- */
 
 import java.io.PrintWriter;
 
@@ -34,7 +33,7 @@ import java.io.PrintWriter;
  * @author Jonas Schaub
  * @version 1.0.0.0
  */
-public interface IPreference extends Comparable<IPreference>, Cloneable {
+public interface IPreference extends Comparable<IPreference> {
     //<editor-fold defaultstate="collapsed" desc="Public properties (get)">
     /**
      * Returns the name of this preference.
@@ -121,16 +120,7 @@ public interface IPreference extends Comparable<IPreference>, Cloneable {
     public int hashCode();
     //</editor-fold>
     //
-    //<editor-fold defaultstate="collapsed" desc="Cloning and copying">
-    /**
-     * Returns a clone of this preference object. When implementing this method the compiler recognizes it as an
-     * override of the clone method in class Object.
-     *
-     * @return a clone of this preference object
-     * @throws CloneNotSupportedException if cloning this specific object fails
-     */
-    public IPreference clone() throws CloneNotSupportedException;
-
+    //<editor-fold defaultstate="collapsed" desc="Copying">
     /**
      * Returns a deep copy of this object.
      *
