@@ -1,29 +1,29 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.preference;
-
-/**
- * TODO:
- * - Implement clone()
- */
 
 import de.unijena.cheminf.mortar.model.util.MiscUtil;
 
@@ -60,7 +60,7 @@ public class SingleNumberPreference extends BasePreference {
     //
     //<editor-fold defaultstate="collapsed" desc="Private class variables">
     /**
-     * Single number content of this preference
+     * Single number content of this preference.
      */
     private double content;
     //</editor-fold>
@@ -101,8 +101,8 @@ public class SingleNumberPreference extends BasePreference {
         if (!BasePreference.isValidName(aName)) {
             throw new IllegalArgumentException("Preference name " + aName + " does not match required pattern!");
         }
-        Double tmpDouble = SingleNumberPreference.parseValidContent(aDoubleString);
-        if (tmpDouble.isNaN()) {
+        double tmpDouble = SingleNumberPreference.parseValidContent(aDoubleString);
+        if (Double.isNaN(tmpDouble)) {
             throw new IllegalArgumentException("The given number " + aDoubleString + " is no valid content!");
         }
         //</editor-fold>
@@ -130,7 +130,7 @@ public class SingleNumberPreference extends BasePreference {
                 //...
                 //break;
                 default:
-                    throw new Exception("Invalid version.");
+                    throw new IOException("Invalid version.");
             }
         } catch (Exception anException) {
             SingleNumberPreference.LOGGER.log(Level.SEVERE, anException.toString(), anException);
@@ -140,6 +140,7 @@ public class SingleNumberPreference extends BasePreference {
     //</editor-fold>
     //
     //<editor-fold defaultstate="collapsed" desc="Public properties (get)">
+
     @Override
     public String getContentRepresentative() {
         return Double.toString(this.content);
@@ -187,8 +188,8 @@ public class SingleNumberPreference extends BasePreference {
      * double is infinite or 'NaN'
      */
     public void setContent(String aDoubleString) throws IllegalArgumentException {
-        Double tmpDouble = SingleNumberPreference.parseValidContent(aDoubleString);
-        if (tmpDouble.isNaN()) {
+        double tmpDouble = SingleNumberPreference.parseValidContent(aDoubleString);
+        if (Double.isNaN(tmpDouble)) {
             throw new IllegalArgumentException("The given number " + aDoubleString + " is no valid content!");
         }
         this.content = tmpDouble;
@@ -196,15 +197,11 @@ public class SingleNumberPreference extends BasePreference {
     //</editor-fold>
     //
     //<editor-fold defaultstate="collapsed" desc="Public methods">
-    @Override
-    public IPreference clone() throws CloneNotSupportedException {
-        throw new UnsupportedOperationException("Not supported yet.");
-    }
 
     @Override
     public SingleNumberPreference copy() {
-        SingleNumberPreference tmpCopy = new SingleNumberPreference(new String(this.name), Double.valueOf(this.content));
-        tmpCopy.guid = new String(this.guid);
+        SingleNumberPreference tmpCopy = new SingleNumberPreference(this.name, this.content);
+        tmpCopy.guid = this.guid;
         return tmpCopy;
     }
 
@@ -230,8 +227,8 @@ public class SingleNumberPreference extends BasePreference {
      * @return true if aContent meets the requirements for contents of this preference
      */
     public static boolean isValidContent(double aContent) {
-        Double tmpDouble = aContent;
-        return !(tmpDouble.isInfinite() || tmpDouble.isNaN());
+        double tmpDouble = aContent;
+        return !(Double.isInfinite(tmpDouble) || Double.isNaN(tmpDouble));
     }
 
     /**
@@ -246,8 +243,8 @@ public class SingleNumberPreference extends BasePreference {
         if (Objects.isNull(aDoubleString) || aDoubleString.isEmpty()) {
             return false;
         }
-        Double tmpResult = SingleNumberPreference.parseValidContent(aDoubleString);
-        return !tmpResult.isNaN();
+        double tmpResult = SingleNumberPreference.parseValidContent(aDoubleString);
+        return !Double.isNaN(tmpResult);
     }
     //</editor-fold>
     //
@@ -255,7 +252,7 @@ public class SingleNumberPreference extends BasePreference {
     /**
      * (Re-)instantiates a new SingleNumberPreference object of version 1.0.0.0 from a line-based text file.
      */
-    private void reloadVersion1000(BufferedReader aReader) throws Exception {
+    private void reloadVersion1000(BufferedReader aReader) throws IOException {
         this.name = aReader.readLine();
         this.guid = aReader.readLine();
         this.content = Double.parseDouble(aReader.readLine());
