@@ -1,24 +1,32 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.model.settings;
+
+import de.unijena.cheminf.mortar.configuration.Configuration;
+import de.unijena.cheminf.mortar.model.io.Exporter;
 
 import javafx.beans.property.Property;
 
@@ -36,6 +44,13 @@ import java.util.Locale;
  */
 public class SettingsContainerTest {
     /**
+     * Constructor to initialize locale and configuration.
+     */
+    public SettingsContainerTest() throws Exception {
+        Locale.setDefault(Locale.of("en", "GB"));
+        Configuration.getInstance();
+    }
+    /**
      * Tests the basic functionalities of SettingsContainer. These are instantiation, restoring default settings,
      * getting the settings, changing the settings, persisting the settings, and reloading them.
      *
@@ -43,26 +58,23 @@ public class SettingsContainerTest {
      */
     @Test
     public void testSettingsContainerBasics() throws Exception {
-        Locale.setDefault(new Locale("en", "GB"));
-        String tmpCsvExportSeparatorTest = ";";
+        Exporter.CSVSeparator tmpCsvExportSeparatorTest = Exporter.CSVSeparator.COMMA;
         //if there is a persisted settings container file already on the machine, it is loaded into the new SettingsContainer object
         SettingsContainer tmpSettingsContainer = new SettingsContainer();
         //restoring to default because a previous settings file with altered settings may have been imported (see below)
         tmpSettingsContainer.restoreDefaultSettings();
-        List<Property> tmpPropertiesList = tmpSettingsContainer.settingsProperties();
-        System.out.println();
-        for (Property tmpProp : tmpPropertiesList) {
-            //recent directory path setting is not included because it is an internal setting
-            System.out.println(tmpProp.getName() + ": " + tmpProp.getValue());
+        List<Property<?>> tmpPropertiesList = tmpSettingsContainer.settingsProperties();
+        for (Property<?> tmpProp : tmpPropertiesList) {
+            Assertions.assertNotNull(tmpProp.getName());
+            Assertions.assertNotNull(tmpProp.getValue());
         }
-        System.out.println(tmpSettingsContainer.recentDirectoryPathSettingProperty().getName() + ": "
-                + tmpSettingsContainer.recentDirectoryPathSettingProperty().getValue());
-        System.out.println();
+        Assertions.assertNotNull(tmpSettingsContainer.recentDirectoryPathSettingProperty().getName());
+        Assertions.assertNotNull(tmpSettingsContainer.recentDirectoryPathSettingProperty().getValue());
         Assertions.assertEquals(SettingsContainer.ROWS_PER_PAGE_SETTING_DEFAULT, tmpSettingsContainer.getRowsPerPageSetting());
         Assertions.assertEquals(SettingsContainer.ADD_IMPLICIT_HYDROGENS_AT_IMPORT_SETTING_DEFAULT, tmpSettingsContainer.getAddImplicitHydrogensAtImportSetting());
         Assertions.assertEquals(tmpSettingsContainer.getNumberOfTasksForFragmentationSettingDefault(), tmpSettingsContainer.getNumberOfTasksForFragmentationSetting());
         Assertions.assertEquals(SettingsContainer.RECENT_DIRECTORY_PATH_SETTING_DEFAULT, tmpSettingsContainer.getRecentDirectoryPathSetting());
-        Assertions.assertEquals(SettingsContainer.KEEP_ATOM_CONTAINER_IN_DATA_MODEL_SETTING_DEFAULT, tmpSettingsContainer.getKeepAtomContainerInDataModelSetting());
+        //Assertions.assertEquals(SettingsContainer.KEEP_ATOM_CONTAINER_IN_DATA_MODEL_SETTING_DEFAULT, tmpSettingsContainer.getKeepAtomContainerInDataModelSetting());
         Assertions.assertEquals(SettingsContainer.ALWAYS_MDLV3000_FORMAT_AT_EXPORT_SETTING_DEFAULT, tmpSettingsContainer.getAlwaysMDLV3000FormatAtExportSetting());
         Assertions.assertEquals(SettingsContainer.CSV_EXPORT_SEPARATOR_SETTING_DEFAULT, tmpSettingsContainer.getCsvExportSeparatorSetting());
         tmpSettingsContainer.setRowsPerPageSetting(SettingsContainer.ROWS_PER_PAGE_SETTING_DEFAULT + 5);
