@@ -1,31 +1,31 @@
 /*
  * MORTAR - MOlecule fRagmenTAtion fRamework
- * Copyright (C) 2023  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
+ * Copyright (C) 2024  Felix Baensch, Jonas Schaub (felix.baensch@w-hs.de, jonas.schaub@uni-jena.de)
  *
  * Source code is available at <https://github.com/FelixBaensch/MORTAR>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 package de.unijena.cheminf.mortar.preference;
 
-/**
- * TODO:
- * - implement comparing test (?)
- * - implement preference name test (?)
- */
-
+import de.unijena.cheminf.mortar.configuration.Configuration;
 import de.unijena.cheminf.mortar.model.util.FileUtil;
 
 import org.junit.jupiter.api.Assertions;
@@ -36,6 +36,7 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.PrintWriter;
+import java.util.Locale;
 
 /**
  * Test class for preferences.
@@ -44,17 +45,18 @@ import java.io.PrintWriter;
  * @version 1.0.0.0
  */
 public class PreferenceTest {
-
     /**
-     * Constructor (empty)
+     * Constructor to initialize locale and configuration.
      */
-    public PreferenceTest() {
+    public PreferenceTest() throws Exception {
+        Locale.setDefault(Locale.of("en", "GB"));
+        Configuration.getInstance();
     }
     //
     /**
      * Tests basic functionalities of class BooleanPreference.
      *
-     * @throws Exception
+     * @throws Exception if anything goes wrong
      */
     @Test
     public void testBooleanPreference() throws Exception {
@@ -67,7 +69,7 @@ public class PreferenceTest {
     /**
      * Tests basic functionalities of class RGBColorPreference.
      *
-     * @throws Exception
+     * @throws Exception if anything goes wrong
      */
     @Test
     public void testColorPreference() throws Exception {
@@ -89,7 +91,7 @@ public class PreferenceTest {
     /**
      * Tests basic functionalities of class SingleIntegerPreference.
      *
-     * @throws Exception
+     * @throws Exception if anything goes wrong
      */
     @Test
     public void testSingleIntegerPreference() throws Exception {
@@ -100,7 +102,7 @@ public class PreferenceTest {
     /**
      * Tests basic functionalities of class SingleNumberPreference.
      *
-     * @throws Exception
+     * @throws Exception if anything goes wrong
      */
     @Test
     public void testSingleNumberPreference() throws Exception {
@@ -111,7 +113,7 @@ public class PreferenceTest {
     /**
      * Tests basic functionalities of class SingleTermPreference.
      *
-     * @throws Exception
+     * @throws Exception if anything goes wrong
      */
     @Test
     public void testSingleTermPreference() throws Exception {
@@ -123,12 +125,11 @@ public class PreferenceTest {
      * Tests basic functionalities of given preference object, like management of public properties and persistence.
      */
     private void testPreferenceBasics(IPreference aPreference) throws Exception {
-        System.out.println();
-        System.out.println(aPreference.getType());
-        System.out.println(aPreference.getContentRepresentative());
-        System.out.println(aPreference.getGUID());
-        System.out.println(aPreference.getName());
-        System.out.println(aPreference.toString());
+        Assertions.assertDoesNotThrow(aPreference::getType);
+        Assertions.assertDoesNotThrow(aPreference::getContentRepresentative);
+        Assertions.assertDoesNotThrow(aPreference::getGUID);
+        Assertions.assertDoesNotThrow(aPreference::getName);
+        Assertions.assertDoesNotThrow(aPreference::toString);
 
         String tmpDir = FileUtil.getAppDirPath() + File.separatorChar + "Test";
         (new File(tmpDir)).mkdirs();
@@ -141,7 +142,7 @@ public class PreferenceTest {
         IPreference tmpPreference = PreferenceFactory.reinitializePreference(tmpReader.readLine(), tmpReader);
         tmpWriter.close();
         tmpReader.close();
-        Assertions.assertTrue(aPreference.getContentRepresentative().equals(tmpPreference.getContentRepresentative()));
+        Assertions.assertEquals(aPreference.getContentRepresentative(), tmpPreference.getContentRepresentative());
         Assertions.assertEquals(aPreference.getName(), tmpPreference.getName());
         Assertions.assertEquals(aPreference.getGUID(), tmpPreference.getGUID());
         Assertions.assertEquals(aPreference.toString(), tmpPreference.toString());
